+++ lib/filesystem/disk/d64.h
@@ -230,7 +230,7 @@ public:
     bool rename(std::string dest) { return false; };
     time_t getLastWrite() override;
     time_t getCreationTime() override;
-    size_t size() override;     
+    uint32_t size() override;     
 
     bool isDir = true;
     bool dirIsOpen = false;
