+++ lib/filesystem/meat_io.h
@@ -75,7 +75,7 @@ public:
     virtual bool rename(std::string dest) = 0;    
     virtual time_t getLastWrite() = 0 ;
     virtual time_t getCreationTime() = 0 ;
-    virtual size_t size() = 0;
+    virtual uint32_t size() = 0;
     virtual uint64_t getAvailableSpace();
 
     virtual bool isText() {
