+++ lib/filesystem/meat_stream.h
@@ -35,9 +35,9 @@ class MStream {
 public:
     virtual ~MStream() {};
 
-    virtual size_t available() = 0;
-    virtual size_t size() = 0;
-    virtual size_t position() = 0;
+    virtual uint32_t available() = 0;
+    virtual uint32_t size() = 0;
+    virtual uint32_t position() = 0;
     virtual size_t error() = 0;
     
     virtual bool isOpen() = 0;
@@ -47,12 +47,12 @@ public:
     virtual void close() = 0;
     virtual bool open() = 0;
 
-    virtual size_t write(const uint8_t *buf, size_t size) = 0;
-    virtual size_t read(uint8_t* buf, size_t size) = 0;
+    virtual uint32_t write(const uint8_t *buf, uint32_t size) = 0;
+    virtual uint32_t read(uint8_t* buf, uint32_t size) = 0;
 
     uint8_t secondaryAddress = 0;
 
-    virtual bool seek(size_t pos, int mode) {
+    virtual bool seek(uint32_t pos, int mode) {
         if(mode == SEEK_SET) {
             return seek(pos);
         }
@@ -63,7 +63,7 @@ public:
             return seek(size() - pos);
         }
     }
-    virtual bool seek(size_t pos) = 0;
+    virtual bool seek(uint32_t pos) = 0;
 
     // For files with a browsable random access directory structure
     // d64, d74, d81, dnp, etc.
