+++ lib/filesystem/tape/tap.h
@@ -101,7 +101,7 @@ public:
     bool rename(std::string dest) { return false; };
     time_t getLastWrite() override { return 0; };
     time_t getCreationTime() override { return 0; };
-    size_t size() override;
+    uint32_t size() override;
 
     bool isDir = true;
     bool dirIsOpen = false;
